requirements.txt
fastapi
uvicorn[standard]
pandas
faiss-cpu
sentence-transformers
openai
scikit-learn
pydantic
python-dotenv